pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isimip-files-auth"
authors = [
    { name = "Jochen Klar", email = "jochen.klar@pik-potsdam.de" },
]
maintainers = [
    { name = "Jochen Klar", email = "jochen.klar@pik-potsdam.de" },
]
description = "A lightweight JWT validation service for the ISIMIP Repository"

readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
classifiers = [
    'Operating System :: OS Independent',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12'
]
dependencies = [
    "Flask~=3.0",
    "gunicorn~=23.0",
    "pyjwt~=2.10",
    "python-dotenv~=1.0",
]
dynamic = ["version"]

[project.scripts]
isimip-files-auth = "isimip_files_auth.cli:main"

[project.urls]
Repository = "https://github.com/ISI-MIP/isimip-files-auth"

[project.optional-dependencies]
dev = [
    "build",
    "pre-commit",
    "pytest",
    "ruff",
    "twine"
]

[tool.setuptools.packages.find]
include = ["isimip_files_auth*"]

[tool.setuptools.package-data]
"*" = ["*"]

[tool.setuptools_scm]
version_file = "isimip_files_auth/_version.py"

[tool.ruff]
target-version = "py311"
select = [
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "PGH",  # pygrep-hooks
    "RUF",  # ruff
    "UP",   # pyupgrade
    "W",    # pycodestyle
    "YTT",  # flake8-2020
]
ignore = [
    "B006",   # mutable-argument-default
    "B007",   # unused-loop-control-variable
    "B018",   # useless-expression
    "RUF012", # mutable-class-default
]

[tool.ruff.isort]
known-first-party = [
    "isimip_files_auth"
]
section-order = [
    "future",
    "standard-library",
    "pytest",
    "flask",
    "third-party",
    "first-party",
    "local-folder"
]

[tool.ruff.isort.sections]
pytest = ["pytest"]
flask = ["flask"]

[tool.pytest.ini_options]
testpaths = ["isimip_files_auth/tests.py"]
